Tie in vehicle logic to findTarget
Helicopters will focus on players outside 30% more unless you're in a stationary vehicle at the time.

Helicopters will also focus more on moving vehicles when seeking targets (every mph/kmph = +1%.

Being in a tile with a tree lowers the chances by 66%. This is on top of being in one of the forest zones.